Code is fixed.
The code was mostly correct. I changed all operands to == and moved the last elif statement to the first if statement. the purpose of the lone sum is to add all three values together except for when 2 or more numbers match. if all three match then it should equal 0. with the all three match statement last it would see that 2 were == and stop executing the function returning the last number that is the same as the other two. it never got to the all equal statement to == 0. putting it first solves the issue.